requirements.txt
python-dotenv
chromadb
llama-index-core
llama-index-readers-file
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
